+++ pokedex/src/components/Pokemones.jsx
@@ -1,4 +1,5 @@
-import { useState, useEffect } from 'react'
+
+import usePokemones from '../hooks/usePokemones'
 import './pokemones.css'
 
 function Pokemon ({id, name, img}){
@@ -17,34 +18,7 @@ function Pokemon ({id, name, img}){
 }
 
 function Pokemones (){
-    const [pokemones, setPokemones]= useState([]);
-
-    useEffect(()=>{
-        const getPokemones = async()=>{
-          const response = await fetch ('https://pokeapi.co/api/v2/pokemon?limit=151&offset=0')
-          const listapokemones = await response.json()
-          const {results} = listapokemones
-          const newPokemones = results.map(async(pokemon)=>{
-            
-            const response = await fetch(pokemon.url)
-            const poke = await response.json()
-    
-            return {
-              id : poke.id,
-              name : poke.name,
-              img: poke.sprites.other.dream_world.front_default
-              
-            }
-          })
-    
-          setPokemones(await Promise.all(newPokemones))
-        }
-      
-        getPokemones();
-        
-      }, [])
-    
-
+    const {pokemones} = usePokemones();
     return (
         <section className='pokemon-container'>
             {
